Add top-level module for Cardano legacy block
