+++ ouroboros-consensus-cardano-legacy-block/src/Legacy/Cardano.hs
@@ -1,0 +1,12 @@
+module Legacy.Cardano (
+    LegacyCardanoBlock
+  , LegacyCardanoEras
+  , LegacyCardanoHardForkConstraints
+  , LegacyCardanoShelleyEras
+  ) where
+
+import           Legacy.Byron.Ledger ()
+import           Legacy.Cardano.Block
+import           Legacy.Cardano.CanHardFork
+import           Legacy.Cardano.Ledger ()
+import           Legacy.Shelley.Ledger ()
